기업 대시보드 및 profile 컴포넌트화
